[CHORE] project folder restructuring
