--- UI/login.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<head>
-	<title>Bucket Lists App | Login Page</title>
-	
-	<link href="http://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet">
-	<link href="assets/plugins/bootstrap/css/bootstrap.min.css" rel="stylesheet" />
-	<link href="assets/css/style.min.css" rel="stylesheet" />
-</head>
-<body>
-	
-	<!-- begin #page-container -->
-	<div id="page-container">
-	    <!-- begin login -->
-        <div class="login login-with-news-feed">
-            <!-- begin news-feed -->
-            <div class="news-feed">
-                <div class="news-image">
-                    <img src="assets/img/bg-7.jpg" alt="" />
-                </div>
-                <div class="news-caption">
-                    <h4 class="caption-title">Bucket Lists App</h4>
-                    <p>
-                        Keep Track of things you want to do, before its too late.
-                    </p>
-                </div>
-            </div>
-            <!-- end news-feed -->
-            <!-- begin right-content -->
-            <div class="right-content">
-                <!-- begin login-header -->
-                <div class="login-header">
-                    <div class="brand">
-                        <span class="logo"></span> Login
-                    </div>
-                </div>
-                <!-- end login-header -->
-                <!-- begin login-content -->
-                <div class="login-content">
-                    <form action="index.html" method="POST" class="margin-bottom-0">
-                        <div class="form-group m-b-15">
-                            <input type="text" class="form-control input-lg" placeholder="Username" />
-                        </div>
-                        <div class="form-group m-b-15">
-                            <input type="text" class="form-control input-lg" placeholder="Password" />
-                        </div>                      
-                        <div class="login-buttons">
-                            <button type="submit" class="btn btn-success btn-block btn-lg">Sign me in</button>
-                        </div>
-                        <div class="m-t-20 m-b-40 p-b-40">
-                            Not a member yet? Click <a href="register.html" class="text-success">here</a> to register.
-                        </div>         
-                    </form>
-                </div>
-                <!-- end login-content -->
-            </div>
-            <!-- end right-container -->
-        </div>
-        <!-- end login -->
-        
-        
-	</div>
-	<!-- end page container -->
-	
-	
-</body>
-</html>
