--- UI/register.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<title>Bucket Lists App | Register Page</title>
-	
-	<!-- ================== BEGIN BASE CSS STYLE ================== -->
-	<link href="http://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet">
-	<link href="assets/plugins/bootstrap/css/bootstrap.min.css" rel="stylesheet" />
-	<link href="assets/css/style.min.css" rel="stylesheet" />
-	<!-- ================== END BASE CSS STYLE ================== -->
-
-</head>
-<body>
-	<!-- begin #page-container -->
-	<div id="page-container">
-	    <!-- begin register -->
-        <div class="register register-with-news-feed">
-            <!-- begin news-feed -->
-            <div class="news-feed">
-                <div class="news-image">
-                    <img src="assets/img/bg-7.jpg" alt="" />
-                </div>
-                <div class="news-caption">
-                    <h4 class="caption-title">Bucket Lists App</h4>
-                    <p>
-                        Keep Track of things you want to do, before its too late.
-                    </p>
-                </div>
-            </div>
-            <!-- end news-feed -->
-            <!-- begin right-content -->
-            <div class="right-content">
-                <!-- begin register-header -->
-                <h1 class="register-header">
-                    Sign Up
-                </h1>
-                <!-- end register-header -->
-                <!-- begin register-content -->
-                <div class="register-content">
-                    <form action="index.html" method="POST" class="margin-bottom-0">                        
-                        <label class="control-label">Username</label>
-                        <div class="row m-b-15">
-                            <div class="col-md-12">
-                                <input type="text" class="form-control" placeholder="Username" />
-                            </div>
-                        </div>
-                        <label class="control-label">Password</label>
-                        <div class="row m-b-15">
-                            <div class="col-md-12">
-                                <input type="text" class="form-control" placeholder="Password" />
-                            </div>
-                        </div>
-                        <label class="control-label">Re-enter Password</label>
-                        <div class="row m-b-15">
-                            <div class="col-md-12">
-                                <input type="text" class="form-control" placeholder="Re-enter Password" />
-                            </div>
-                        </div>
-                        <div class="register-buttons">
-                            <button type="submit" class="btn btn-primary btn-block btn-lg">Sign Up</button>
-                        </div>
-                        <div class="m-t-20 m-b-40 p-b-40">
-                            Already a member? Click <a href="login.html">here</a> to login.
-                        </div>
-
-                    </form>
-                </div>
-                <!-- end register-content -->
-            </div>
-            <!-- end right-content -->
-        </div>
-        <!-- end register -->
-        
-    </div>
-	<!-- end page container -->
-</body>
-</html>
