--- UI/tasks.html
@@ -1,176 +1,0 @@
-<!DOCTYPE html>
-<head>
-	<title>Bucket-List App | Tasks</title>
-	<link href="http://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet">
-	<link href="assets/plugins/bootstrap/css/bootstrap.min.css" rel="stylesheet" />
-	<link href="assets/css/style.min.css" rel="stylesheet" />
-</head>
-<body>	
-	<div id="page-container" class="page-without-sidebar page-header-fixed">
-		<div id="header" class="header navbar navbar-default navbar-fixed-top">
-			<div class="container-fluid">
-				<div class="navbar-header">
-					<a href="bucketlists.html" class="navbar-brand"><span class="navbar-logo"></span>Bucket List App</a>
-				</div>
-				<!-- Logout Link -->
-				<ul class="nav navbar-nav navbar-right">
-					<li class="dropdown navbar-user">
-						<a href="javascript:;" class="dropdown-toggle" data-toggle="dropdown">
-							<span class="hidden-xs">Logout</span></b>
-						</a>
-					</li>
-				</ul>
-				<!-- end logout link -->	
-			</div>
-		</div>
-		
-		<!-- begin #content -->
-		<div id="content" class="container">
-			<!-- begin page-header -->
-			<h6 class="page-header"></h6>
-			<!-- end page-header -->
-			<div class="panel panel-inverse">
-			    <div class="panel-heading">
-					<div class="panel-heading-btn">
-						<a href="#modal-dialog" class="btn btn-sm btn-success" data-toggle="modal">New Task</a>
-					</div>
-			        <h2 class="panel-title">Tasks</h2>
-			    </div>
-				<!-- new task modal -->
-				<div class="modal fade" id="modal-dialog">
-					<div class="modal-dialog">
-						<div class="modal-content">
-							<div class="modal-header">
-								<button type="button" class="close" data-dismiss="modal" aria-hidden="true">×</button>
-								<h4 class="modal-title">New Task</h4>
-							</div>
-							<div class="modal-body">
-								<form class="form-horizontal" action="/" method="POST">
-									<fieldset>
-										<div class="form-group">
-											<div class="col-md-12">
-												<input type="email" class="form-control" placeholder="Enter Task" />
-											</div>
-										</div>                                 
-									</fieldset>
-                            	</form>
-							</div>
-								<div class="modal-footer">
-									<a href="javascript:;" class="btn btn-sm btn-white" data-dismiss="modal">Close</a>
-									<a href="javascript:;" class="btn btn-sm btn-success">Add Task</a>
-								</div>
-							</div>
-						</div>
-					</div>
-				<!-- end new task modal -->
-				
-			    <div class="panel-body">
-					<!-- all tasks table -->
-                        <div class="panel-body">
-							<div class="table-responsive">
-								<table id="data-table" class="table table-striped table-bordered" width="100%">
-								<colgroup>
-									<col span="1" style="width: 10%;">
-									<col span="1" style="width: 60%;">
-                                    <col span="1" style="width: 10%;">
-									<col span="1" style="width: 10%;">
-									<col span="1" style="width: 10%;">
-								</colgroup>
-									<thead>
-										<tr>
-											<th>#</th>
-											<th>Task</th>
-											<th>complete</th>
-											<th>Edit</th>
-											<th>Delete</th>
-										</tr>
-									</thead>
-									<tbody>
-										<tr>
-											<td>1</td>
-											<td>Task 1</td>
-											<td class="text-center input-lg"><input type="checkbox" value="" /></td>
-											<td class="text-center"><button type="button" class="btn btn-info" data-toggle="modal" data-target="#modal1"><span class="glyphicon glyphicon-edit"></span></button> </td>
-											<td class="text-center"><button type="button" class="btn btn-info" data-toggle="modal" data-target="#modal2"><span class="glyphicon glyphicon-remove"></span></button> </td>																		
-										</tr>
-										<tr>
-											<td>2</td>
-											<td>Task 2</td>
-											<td class="text-center input-lg"><input type="checkbox" value="" /></td>
-											<td class="text-center"><button type="button" class="btn btn-info" data-toggle="modal" data-target="#modal1"><span class="glyphicon glyphicon-edit"></span></button> </td>
-											<td class="text-center"><button type="button" class="btn btn-info" data-toggle="modal" data-target="#modal2"><span class="glyphicon glyphicon-remove"></span></button> </td>																		
-										</tr>
-										<tr>
-											<td>3</td>
-											<td>Task 3</td>
-											<td class="text-center input-lg"><input type="checkbox" value="" /></td>
-											<td class="text-center"><button type="button" class="btn btn-info" data-toggle="modal" data-target="#modal1"><span class="glyphicon glyphicon-edit"></span></button> </td>
-											<td class="text-center"><button type="button" class="btn btn-info" data-toggle="modal" data-target="#modal2"><span class="glyphicon glyphicon-remove"></span></button> </td>																		
-										</tr>
-									</tbody>
-								</table>
-							</div>
-						</div>
-						<!-- end of all tasks table -->
-
-						<!-- Modal For Editing a Task -->
-						<div class="modal fade" id="modal1">
-							<div class="modal-dialog">
-								<div class="modal-content">
-									<div class="modal-header">
-										<button type="button" class="close" data-dismiss="modal" aria-hidden="true">×</button>
-										<h4 class="modal-title">Edit Task</h4>
-									</div>
-									<div class="modal-body">
-										<form class="form-horizontal" action="/" method="POST">
-										<fieldset>
-											<div class="form-group">
-												<div class="col-md-12">
-													<input type="email" class="form-control" placeholder="Task Name" />
-												</div>
-											</div>                                 
-											
-										</fieldset>
-									</form>
-									</div>
-									<div class="modal-footer">
-										<a href="javascript:;" class="btn btn-sm btn-white" data-dismiss="modal">Close</a>
-										<a href="javascript:;" class="btn btn-sm btn-success">Edit Task</a>
-									</div>
-								</div>
-							</div>
-						</div>
-						<!-- End of edit task modal -->
-						
-						<!-- Modal For Removing a task confirmation -->
-						<div class="modal fade" id="modal2">
-							<div class="modal-dialog">
-								<div class="modal-content">
-									<div class="modal-header">
-										<button type="button" class="close" data-dismiss="modal" aria-hidden="true">×</button>
-										<h4 class="modal-title">Remove Task</h4>
-									</div>
-									<div class="modal-body">
-										<h5>Are you sure to remove task?
-									</div>
-									<div class="modal-footer">
-										<a href="javascript:;" class="btn btn-sm btn-white" data-dismiss="modal">Close</a>
-										<a href="javascript:;" class="btn btn-sm btn-success">Remove Task</a>
-									</div>
-								</div>
-							</div>
-						</div>
-						<!-- End of remove task confirmation modal -->
-						
-					</div>
-			        </p>
-			    </div>
-			</div>   
-		</div>		
-	</div>	
-
-	<script src="assets/plugins/jquery/jquery-1.9.1.min.js"></script>
-	<script src="assets/plugins/bootstrap/js/bootstrap.min.js"></script>
-	
-</body>
-</html>
