+++ app/static/plugins/bootstrap/fonts/glyphicons-halflings-regular.woff
